refactor, docstrings, and version bump
