+++ pyproject.toml
@@ -15,7 +15,7 @@ allow-direct-references = true
 # https://peps.python.org/pep-0621/
 [project]
 name = "opm-processing-v2"
-version = "0.3.0"
+version = "0.4.0"
 description = "qi2lab OPM post-processing package v2"
 readme = "README.md"
 requires-python = ">=3.12"
