+++ src/opm_processing/__init__.py
@@ -1,5 +1,5 @@
 """qi2lab OPM post-processing package v2"""
 
-__version__ = '0.3.0'
+__version__ = '0.4.0'
 __author__ = "Douglas Shepherd"
 __email__ = "douglas.shepherd@asu.edu"
