+++ src/opm_processing/dataio/zarr_handlers.py
@@ -10,6 +10,7 @@ History:
 
 from pathlib import Path
 import tensorstore as ts
+import numpy as np
 
 def create_multiscale_dict(
     axes: dict, 
@@ -241,3 +242,34 @@ def write_via_tensorstore(ts_store, data, data_location):
     # Initiate the write operation (this returns a Future).
     future = ts_store[index].write(data)
     return future
+
+
+class TensorStoreWrapper:
+    """Wrapper for tensorstore array to provide ndarray properties.
+    
+    Parameters
+    ----------
+    ts_array: tensorstore
+        tensorstore array
+    """
+    
+    def __init__(self, ts_array):
+        self.ts_array = ts_array
+        self.shape = tuple(ts_array.shape)
+        self.dtype = ts_array.dtype.numpy_dtype
+        self.ndim = len(self.shape)
+
+    def __getitem__(self, idx):
+        """Return item from tensorstore array at requested indices.
+        
+        Parameters
+        ----------
+        idx: list
+            slice indices
+        """
+        
+        return self.ts_array[idx].read().result()
+
+    def __array__(self):
+        """Return fake array with correct dtype."""
+        return np.empty(self.shape, dtype=self.dtype)
